fix(XML Binding): fix gml geometry parsing regarding srs dimension. Add other improvments for GML to JTS conversion.
GML geometries should define a srsDimension attribute to specify its dimension (2D or 3D). The problem is that sometimes, this attribute is on a parent markup, so we cannot access it directly from the geometry.

+++ modules/jaxb-xml-binding/geotk-xml-gml/src/main/java/org/geotoolkit/gml/xml/v311/AbstractGeometryType.java
@@ -323,7 +323,7 @@ public abstract class AbstractGeometryType extends AbstractGMLType implements Ge
         if(bi == null){
             return 2;
         }else{
-            return bi.intValue();
+            return bi;
         }
     }
